refactor: personalization projection을 최상위 패키지 infrastructure로 이동
