+++ src/main/java/com/techfork/personalization/infrastructure/PersonalizationProfileDocument.java
@@ -1,6 +1,5 @@
-package com.techfork.domain.personalization.document;
+package com.techfork.personalization.infrastructure;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AccessLevel;
 import lombok.Builder;
