+++ src/test/java/com/techfork/domain/search/service/SearchServiceImplTest.java
@@ -6,8 +6,8 @@ import co.elastic.clients.elasticsearch.core.SearchResponse;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.util.ObjectBuilder;
 import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
-import com.techfork.domain.personalization.document.PersonalizationProfileDocument;
-import com.techfork.domain.personalization.repository.PersonalizationProfileDocumentRepository;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocumentRepository;
 import com.techfork.post.domain.Post;
 import com.techfork.post.domain.projection.PostDocument;
 import com.techfork.post.infrastructure.PostRepository;
