+++ src/test/java/com/techfork/evaluation/recommendation/RecommendationEvaluationService.java
@@ -17,9 +17,9 @@ import com.techfork.domain.recommendation.service.MmrService;
 import com.techfork.domain.recommendation.service.MmrService.MmrCandidate;
 import com.techfork.domain.recommendation.service.MmrService.MmrResult;
 import com.techfork.global.util.RrfScorer;
-import com.techfork.domain.personalization.document.PersonalizationProfileDocument;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.useraccount.domain.User;
-import com.techfork.domain.personalization.repository.PersonalizationProfileDocumentRepository;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocumentRepository;
 import com.techfork.global.elasticsearch.query.VectorQueryBuilder;
 import com.techfork.global.util.TimeDecayStrategy;
 import lombok.extern.slf4j.Slf4j;
