+++ src/test/java/com/techfork/evaluation/recommendation/setup/UserDataSetupAndExporter.java
@@ -7,10 +7,10 @@ import com.techfork.evaluation.recommendation.setup.components.FileExporter;
 import com.techfork.evaluation.recommendation.setup.components.TestDataGenerator;
 import com.techfork.evaluation.recommendation.setup.components.TestDataGenerator.UserCreationResult;
 import com.techfork.evaluation.recommendation.util.EvaluationFixtureLoader;
-import com.techfork.domain.personalization.document.PersonalizationProfileDocument;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.useraccount.domain.User;
 import com.techfork.useraccount.domain.enums.EInterestCategory;
-import com.techfork.domain.personalization.repository.PersonalizationProfileDocumentRepository;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocumentRepository;
 import com.techfork.useraccount.infrastructure.UserRepository;
 import com.techfork.global.common.IntegrationTestBase;
 import lombok.extern.slf4j.Slf4j;
