+++ src/test/java/com/techfork/evaluation/recommendation/setup/components/GroundTruthGenerator.java
@@ -3,7 +3,7 @@ package com.techfork.evaluation.recommendation.setup.components;
 import com.techfork.post.domain.projection.PostDocument;
 import com.techfork.post.domain.Post;
 import com.techfork.post.infrastructure.PostDocumentRepository;
-import com.techfork.domain.personalization.document.PersonalizationProfileDocument;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.global.llm.LlmClient;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
