+++ src/test/java/com/techfork/evaluation/recommendation/setup/components/TestDataGenerator.java
@@ -2,10 +2,10 @@ package com.techfork.evaluation.recommendation.setup.components;
 
 import com.techfork.post.domain.Post;
 import com.techfork.post.infrastructure.PostRepository;
-import com.techfork.domain.personalization.document.PersonalizationProfileDocument;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.useraccount.domain.User;
 import com.techfork.useraccount.domain.enums.EInterestCategory;
-import com.techfork.domain.personalization.repository.PersonalizationProfileDocumentRepository;
+import com.techfork.personalization.infrastructure.PersonalizationProfileDocumentRepository;
 import com.techfork.domain.personalization.service.PersonalizationProfileService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
